Create iml-warp-drive-service
Currently for each stateful object requested via the API,
we tack on extra information regarding it's available jobs +
transitions, and locks.

This is done in:

https://github.com/whamcloud/integrated-manager-for-lustre/blob/5f8dcc3615388527c833df348617119352f51d07/chroma_api/utils.py#L198-L280

This is primarily done for convienience, but adds a bunch of overhead
both in RPC calls and database access.

Instead of fetching locks on demand, we can move to a push model.

Add a `locks` queue to rabbit that will push `LockChange` events from
the job scheduler.

Then create a new service `iml-warp-drive` that will consume this queue
and push the current lock state using Server-Sent Events (SSE) to any
interested clients.

Wire this service into Nginx so it can be accessed over the same URL +
port as everything else.

This service is both written in Rust and colocated with other services
in the IML manager.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>