Deepcopy steps when passing to RunJobThread
We are passing full model instances between threads.

This is leading to race conditions, which can be seen while trying to
clone an instance within a thread.

We should be passing immutable types between threads, but we currently
require an instance within the threads for `job_scheduler_notify` calls.

As as start, `deepcopy` the model instance *before* it enters a thread and
becomes subject to "parallelism".

Signed-off-by: Joe Grund <jgrund@whamcloud.io>